# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Customer Care API",
        "description": "API for Customer Care System",
        "version": "1.0.0"
    },
    "paths": {
        "/api/tickets/{id}/responses": {
            "get": {
                "tags": [
                    "Responses"
                ],
                "summary": "Get all responses for a ticket",
                "operationId": "642d4ddfdd8958f444606da41f43ae6e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ticket ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of responses"
                    },
                    "404": {
                        "description": "Ticket not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/response": {
            "post": {
                "tags": [
                    "Responses"
                ],
                "summary": "Add a response to a ticket",
                "operationId": "44989cf081dbea4be61c494f57ed7c98",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Response added"
                    },
                    "404": {
                        "description": "Ticket not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/response/{response}": {
            "get": {
                "tags": [
                    "Responses"
                ],
                "summary": "Get response details",
                "operationId": "a91a37662e96fc8190e3ea0672dd28d0",
                "parameters": [
                    {
                        "name": "response",
                        "in": "path",
                        "description": "Response ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response details"
                    },
                    "404": {
                        "description": "Response not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Responses"
                ],
                "summary": "Update a response",
                "operationId": "e6736e0f118b41e389ebdaa043cfa5cb",
                "parameters": [
                    {
                        "name": "response",
                        "in": "path",
                        "description": "Response ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Response updated"
                    },
                    "404": {
                        "description": "Response not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Responses"
                ],
                "summary": "Delete a response",
                "operationId": "979ce8a79db9d398457771dac78f7331",
                "parameters": [
                    {
                        "name": "response",
                        "in": "path",
                        "description": "Response ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response deleted"
                    },
                    "404": {
                        "description": "Response not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}